requirements.txt
streamlit
pandas
openpyxl
XlsxWriter
python-barcode
Pillow